moved .h for swig C#
